backend/requirements/base.txt
asyncpg
celery[redis]
fastapi-users[redis,oauth]
fastapi_users_db_sqlmodel
python-dotenv
nest_asyncio
sqlmodel